chore: revert to commit 4e3418b7d31aea172a27743f654de934221ff717

+++ backend/app/src/main/java/org/scrollSystem/service/ScrollService.java
@@ -12,6 +12,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.sql.Timestamp;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -71,9 +74,48 @@ public class ScrollService {
         if (optionalFileStorage.isEmpty()) {
             throw new ValidationException("File id is not exists");
         }
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Integer user_id = user.getId();
 
         FileStorage fileStorage = optionalFileStorage.get();
+
+        if (!fileStorage.getOwner().getId().equals(user_id) && user.getRole().equals("ROLE_USER")) {
+            throw new ValidationException("You are not the owner of the file so you can not delete this scroll");
+        }
+
         s3Service.deleteFile(fileStorage.getTitle());
         fileRepository.delete(fileStorage);
+    }
+
+
+    public List<FileResponse> getScrolls() {
+        List<FileResponse> response = new ArrayList<>();
+
+        List<FileStorage> scrollsList = fileRepository.findAll();
+        for (FileStorage fileStorage: scrollsList) {
+            FileResponse scrollResponse = FileResponse.builder()
+                    .fileId(fileStorage.getFileId())
+                    .title(fileStorage.getTitle())
+                    .filePath(fileStorage.getFilePath())
+                    .fileSize(fileStorage.getFileSize())
+                    .fileType(fileStorage.getFileType())
+                    .uploadDate(fileStorage.getUploadDate())
+                    .downloadAmount(fileStorage.getDownloadAmount())
+                    .build();
+            User owner = fileStorage.getOwner();
+            UserResponse userResponse = UserResponse.builder()
+                    .id(owner.getId())
+                    .firstName(owner.getFirstName())
+                    .lastName(owner.getLastName())
+                    .email(owner.getEmail())
+                    .username(owner.getUsername())
+                    .phone(owner.getPhone())
+                    .role(owner.getRole())
+                    .build();
+
+            scrollResponse.setOwner(userResponse);
+            response.add(scrollResponse);
+        }
+        return response;
     }
 }
